requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
django_extensions